# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wallet Management APIs",
        "description": "API for managing wallets and transfers with double-entry bookkeeping",
        "contact": {
            "email": "samspike46@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/wallets": {
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Create a new wallet for a user",
                "description": "Initialize a wallet for a user. A user may only have one wallet.",
                "operationId": "createWallet",
                "requestBody": {
                    "description": "Wallet creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "initial_balance": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wallet created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict - User already has a wallet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "tokenAuth": []
                    }
                ]
            }
        },
        "/api/v1/wallets/{id}": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Get wallet details and balance",
                "description": "Fetch wallet details including balance and transaction summary (credit, debit totals).",
                "operationId": "getWallet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wallet ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "tokenAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Delete a wallet",
                "description": "Delete a wallet. Only allowed if wallet balance is zero.",
                "operationId": "deleteWallet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wallet ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Wallet balance is not zero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "tokenAuth": []
                    }
                ]
            }
        },
        "/api/v1/wallets/{id}/fund": {
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Add funds to wallet",
                "description": "Fund wallet by adding credit transaction. Creates a wallet transaction of type credit.",
                "operationId": "fundWallet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wallet ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fund wallet data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 1,
                                        "example": 500
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Bank transfer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wallet funded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "tokenAuth": []
                    }
                ]
            }
        },
        "/api/v1/wallets/{id}/withdraw": {
            "post": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Withdraw funds from wallet",
                "description": "Withdraw funds from wallet. Creates a wallet transaction of type debit. Must enforce balance > withdrawal amount.",
                "operationId": "withdrawFromWallet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wallet ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Withdraw wallet data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 1,
                                        "example": 200
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Cash withdrawal",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Withdrawal successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Insufficient balance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wallet not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "tokenAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "description": "Standard API response format",
                "required": [
                    "status_code",
                    "message"
                ],
                "properties": {
                    "status_code": {
                        "description": "HTTP status code",
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "description": "Response message",
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "description": "Response data",
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    },
                    "errors": {
                        "description": "Array of error messages",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "description": "Error response format",
                "required": [
                    "status_code",
                    "message"
                ],
                "properties": {
                    "status_code": {
                        "description": "HTTP status code",
                        "type": "integer",
                        "example": 400
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "description": "Array of validation errors",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "The amount field is required."
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "tokenAuth": {
                "type": "apiKey",
                "description": "Use token: VG@123 in header",
                "name": "token",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "tokenAuth": []
        }
    ],
...
}